+++ pkg.nix
@@ -1,6 +1,7 @@
 { mkDerivation, stdenv, ghc
 , base, bytestring, constraints, tasty, tasty-hunit, tasty-quickcheck
 
+
 # Optional dependencies
 , aeson ? null
 , binary ? null
@@ -9,7 +10,11 @@
 , hashable ? null
 , serialise ? null
 , store ? null
+, xmlbf ? null
+, text ? null
 }:
+
+assert !(isNull xmlbf) -> !(isNull text);
 
 let
 extraDeps =
@@ -19,7 +24,8 @@ extraDeps =
   stdenv.lib.optionals (!(isNull deepseq)) [ deepseq ] ++
   stdenv.lib.optionals (!(isNull hashable)) [ hashable ] ++
   stdenv.lib.optionals (!(isNull serialise)) [ serialise ] ++
-  stdenv.lib.optionals (!(isNull store)) [ store ];
+  stdenv.lib.optionals (!(isNull store)) [ store ] ++
+  stdenv.lib.optionals (!(isNull xmlbf)) [ xmlbf text ];
 
 in mkDerivation rec {
   pname = "safe-money";
@@ -38,5 +44,6 @@ in mkDerivation rec {
     stdenv.lib.optionals (isNull deepseq) [ "-f-deepseq" ] ++
     stdenv.lib.optionals (isNull hashable) [ "-f-hashable" ] ++
     stdenv.lib.optionals (isNull serialise) [ "-f-serialise" ] ++
-    stdenv.lib.optionals (isNull store) [ "-f-store" ];
+    stdenv.lib.optionals (isNull store) [ "-f-store" ] ++
+    stdenv.lib.optionals (isNull xmlbf) [ "-f-xmlbf" ];
 }